Added UpdateChecker and fixed scoreboard - Added UpdateChecker to check for updates - Fixed scoreboard to properly display first, second and third place - Cleaned up code and made ready for BETA release

+++ src/main/java/me/koenn/purge/Purge.java
@@ -1,12 +1,14 @@
 package me.koenn.purge;
 
 import me.koenn.purge.commands.PurgeCommand;
+import me.koenn.purge.listeners.JoinListener;
+import me.koenn.purge.listeners.KillListener;
 import me.koenn.purge.util.ConfigManager;
+import me.koenn.purge.util.References;
+import me.koenn.purge.util.UpdateChecker;
 import org.bukkit.Bukkit;
+import org.bukkit.permissions.ServerOperator;
 import org.bukkit.plugin.java.JavaPlugin;
-
-import java.text.SimpleDateFormat;
-import java.util.Date;
 
 /**
  * A Minecraft PvP purge plugin
@@ -22,6 +24,7 @@ public final class Purge extends JavaPlugin {
 
     private static ConfigManager configManager;
     private static Purge instance;
+    private static References references;
 
     public static ConfigManager getConfigManager() {
         return configManager;
@@ -31,20 +34,38 @@ public final class Purge extends JavaPlugin {
         return instance;
     }
 
+    public static References getReferences() {
+        return references;
+    }
+
     @Override
     public void onEnable() {
         instance = this;
 
         this.getLogger().info("All credits for this plugin go to Koenn");
 
+        configManager = new ConfigManager(this);
+        references = new References();
+        if (!references.load()) {
+            Purge.getInstance().getLogger().severe("Error while loading configurations!");
+            Purge.getInstance().getLogger().severe("The plugin will now disable, please fix any mistake in the config.yml file!");
+            Bukkit.getPluginManager().disablePlugin(this);
+            return;
+        }
+
         this.getCommand("purge").setExecutor(new PurgeCommand());
 
-        configManager = new ConfigManager(this);
+        Bukkit.getPluginManager().registerEvents(new KillListener(), this);
+        Bukkit.getPluginManager().registerEvents(new JoinListener(), this);
 
-        Bukkit.getScheduler().scheduleSyncRepeatingTask(this, () -> {
-            Date now = new Date();
-            SimpleDateFormat format = new SimpleDateFormat("");
-        }, 100, 1200);
+        Bukkit.getScheduler().scheduleSyncDelayedTask(this, () -> {
+            if (UpdateChecker.updateAvailable()) {
+                this.getLogger().info("An update is available! Download it here: https://github.com/Koennn/Purge");
+                Bukkit.getOnlinePlayers().stream().filter(ServerOperator::isOp).forEach(UpdateChecker::sendUpdateMessage);
+            }
+        }, 40);
+
+        this.getLogger().info("The plugin has been successfully setup, and ready to start the Purge!");
     }
 
     @Override
